重整订单的代码逻辑，添加了一个文字效果，实现了经验条基础动效

+++ Assets/Scripts/MainScene/GameLoop/GameLoopData.cs
@@ -14,7 +14,7 @@ namespace YogurtCulture.GameLoop
 
         public int todaySuccessOrders = 0;
         public int todayFailOrders = 0;
-        public float todayReputationGain = 0f;
+        public int todayReputationGain = 0;
         public int todayLevelUps = 0;
 
         public event System.Action<int, bool> OnOrderCompleted;
